move all components into a ui folder
