+++ client/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { GamePageComponent } from './game/containers/game-page/game-page.component';
-import { IntroPageComponent } from './intro/containers/intro-page/intro-page.component';
-import { LobbyPageComponent } from './lobby/containers/lobby-page/lobby-page.component';
-import { NotFoundPageComponent } from './shared/containers/not-found-page/not-found-page.component';
+import { GamePageComponent } from './ui/game/containers/game-page/game-page.component';
+import { IntroPageComponent } from './ui/intro/containers/intro-page/intro-page.component';
+import { LobbyPageComponent } from './ui/lobby/containers/lobby-page/lobby-page.component';
+import { NotFoundPageComponent } from './ui/shared/containers/not-found-page/not-found-page.component';
 
 const routes: Routes = [
   {
