+++ client/src/app/data/types/Suit.ts
