+++ client/src/app/ui/game/components/board/board.component.html
