+++ client/src/app/ui/game/components/board/board.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../variables' as vars;
+@use '../../../../../variables' as vars;
 
 :host {
   display: grid;
