+++ client/src/app/ui/game/components/card/card.component.html
