+++ client/src/app/ui/intro/containers/intro-page/intro-page.component.scss
