+++ client/src/app/ui/intro/containers/options/options.component.scss
