+++ client/src/app/ui/lobby/containers/lobby-page/lobby-page.component.scss
